Retry connecting on ECONNREFUSED for ten seconds
If both mpd and krunner-mpd are started around the same time (e.g. on
system startup), it makes sense to wait for mpd to appear so that the
user does not have to worry too much about creating a proper dependency.
Thus, this commit makes krunner-mpd retry ten times if ECONNREFUSED is
received, waiting one second between each attempt.